testing inventory

+++ Inventory-management/category/category_service.py
@@ -23,11 +23,28 @@ def get_category(db: Session, category_id: int):
 def list_categories(db: Session):
     return repository.get_all(db)
 
+
+from item.item_models import Item
+
 def delete_category(db: Session, category_id: int):
     category = repository.get_by_id(db, category_id)
     if not category:
-        raise HTTPException(status_code=404, detail="Category not found")
+        raise HTTPException(404, "Category not found")
+
+    # 🔴 check if category is used
+    item_exists = db.query(Item).filter(
+        Item.category_id == category_id
+    ).first()
+
+    if item_exists:
+        raise HTTPException(
+            status_code=400,
+            detail="Cannot delete category. Items exist under this category."
+        )
+
     repository.delete(db, category)
+    return {"detail": "Category deleted successfully"}
+
     
 def update_category(db: Session, category_id: int, updated_data):
     category = repository.get_by_id(db, category_id)
